requirements.txt
matplotlib==3.1.3
numpy==1.18.1
PyYAML==5.3
tweepy==3.8.0